requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langchain-google-genai==2.0.10
google-generativeai
google-genai
faiss-cpu
fastapi
uvicorn
streamlit
python-dotenv
fpdf2
pypdf
tavily-python